requirements.txt
jsonschema-spec
requests==2.25.1
urllib3==1.26.7
chardet==4.0.0
charset_normalizer==2.0.12
python-dotenv==0.19.0